Wrap the `Configuration` in an `RwLock` instead (#17)
This changes the `Mutex` wrapper around `Configuration` inside of the `Framework` to `RwLock`. With `Mutex`, there are high risks of encountering deadlocks if one were to read/modify the configuration inside of a command, as many locks could be held across tasks. `RwLock` reduces these risks, allowing consumers to safely read and modify the configuration.