Reduce instances of defaulted generic parameters
For convenience, we default generic parameters related to the data and
error type to `DefaultData` and `DefaultError`. However, this was done
excessively. It is not necessary to default the parameters on every
type/alias. The overuse had even caused a bug in the definition of
hooks, which prevented using hooks with data and error type different
from `DefaultData` and `DefaultError`.